+++ tests/test_upload.py
@@ -20,7 +20,7 @@ import pretend
 import pytest
 
 from twine.commands import upload
-from twine import package, cli
+from twine import package, cli, exceptions
 import twine
 
 import helpers
@@ -93,6 +93,40 @@ def test_get_config_old_format(tmpdir):
             "more info: "
             "https://docs.python.org/distutils/packageindex.html#pypirc\n"
         ).format(pypirc)
+
+
+def test_deprecated_repo(tmpdir):
+    with pytest.raises(exceptions.UploadToDeprecatedPyPIDetected) as err:
+        pypirc = os.path.join(str(tmpdir), ".pypirc")
+        dists = ["tests/fixtures/twine-1.5.0-py2.py3-none-any.whl"]
+
+        with open(pypirc, "w") as fp:
+            fp.write(textwrap.dedent("""
+                [pypi]
+                repository: https://pypi.python.org/pypi/
+                username:foo
+                password:bar
+            """))
+
+        upload.upload(dists=dists, repository="pypi", sign=None, identity=None,
+                      username=None, password=None, comment=None,
+                      cert=None, client_cert=None,
+                      sign_with=None, config_file=pypirc, skip_existing=False,
+                      repository_url=None,
+                      )
+
+        assert err.args[0] == (
+            "You're trying to upload to the legacy PyPI site "
+            "'https://pypi.python.org/pypi/'. "
+            "Uploading to those sites is deprecated. \n "
+            "The new sites are pypi.org and test.pypi.org. Try using "
+            "https://upload.pypi.org/legacy/ "
+            "(or https://test.pypi.org/legacy/) "
+            "to upload your packages instead. "
+            "These are the default URLs for Twine now. \n "
+            "More at "
+            "https://packaging.python.org/guides/migrating-to-pypi-org/ ."
+            )
 
 
 def test_skip_existing_skips_files_already_on_PyPI(monkeypatch):